Quantity in Item is Double.

+++ backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/endpoint/dto/ItemDto.java
@@ -26,10 +26,10 @@ public record ItemDto(
     String brand,
     @NotNull(message = "The actual quantity cannot be empty")
     @Min(value = 0, message = "The actual quantity must be positive")
-    Long quantityCurrent,
+    Double quantityCurrent,
     @NotNull(message = "The total quantity cannot be empty")
     @Min(value = 0, message = "The total quantity must be positive")
-    Long quantityTotal,
+    Double quantityTotal,
     @NotNull(message = "The unit cannot be null")
     UnitDto unit,
     @FutureOrPresent(message = "You cannot store products which are over the expire date")
@@ -124,7 +124,7 @@ public record ItemDto(
             + '}';
     }
 
-    public ItemDto withUpdatedQuantity(Long updatedQuantity) {
+    public ItemDto withUpdatedQuantity(Double updatedQuantity) {
         return new ItemDto(
             itemId,
             ean,
